👌 Resolve review issue
